Authorization with Policies

+++ app/Http/Controllers/Thread/ThreadController.php
@@ -124,12 +124,13 @@ class ThreadController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-
+        $this->authorize('update', $thread);
+        
         $thread->delete();
 
         if (request()->wantsJson()) {
             
-            return response([], 204);
+            return response([], 403);
         
         }
